Dynamic endpoints handlers
chttpd hardcoded handlers are replaced with with dynamic url handlers.

This is a special branch that pull the prepared branches (of the same name)
of chttpd, mem3, global_changes, setup_httpd, mango_httpd, which have the
relevant config files.

See chttpd/27037-5-dynamic-endpoint-handlers.

This is a re-do of PR 10 for bitrot:
https://github.com/apache/couchdb-chttpd/pull/10

BugzID: 27037